♻️ Adding fullWidth prop
