+++ libraries/core-react/src/components/Button/Button.tsx
@@ -14,6 +14,7 @@ import {
   outlineTemplate,
   spacingsTemplate,
 } from '../../utils'
+import { ButtonFullWidth } from './ButtonFullWidth'
 
 type Colors = 'primary' | 'secondary' | 'danger'
 type Variants = 'contained' | 'outlined' | 'ghost' | 'ghost_icon'
@@ -47,44 +48,12 @@ const getToken = (variant: Variants, color: Colors): ButtonToken => {
   }
 }
 
-const ButtonCenterContent = styled.span`
-  text-align: center;
-  flex: 1;
-`
-
 const ButtonInner = styled.span`
-  height: 100%;
-  display: flex;
+  display: grid;
+  grid-gap: 8px;
+  grid-auto-flow: column;
   align-items: center;
-
-  > img:first-child,
-  > svg:first-child {
-    margin-right: 8px;
-  }
-
-  > img:last-child,
-  > svg:last-child {
-    margin-left: 8px;
-  }
-
-  > img:only-child,
-  > svg:only-child {
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  > span:first-child {
-    margin-left: 32px;
-  }
-
-  > span:last-child {
-    margin-right: 32px;
-  }
-
-  > span:only-child {
-    margin-right: 0;
-    margin-left: 0;
-  }
+  height: 100%;
 `
 
 const Base = ({ token }: { token: ButtonToken }) => {
@@ -179,6 +148,8 @@ export type ButtonProps = {
    * @default 'button'
    */
   type?: string
+  /** FullWidth (stretched) button  */
+  fullWidth?: boolean
 } & (
   | ButtonHTMLAttributes<HTMLButtonElement>
   | AnchorHTMLAttributes<HTMLAnchorElement>
@@ -195,6 +166,7 @@ export const Button = forwardRef<
     disabled = false,
     href,
     tabIndex = 0,
+    fullWidth = false,
     ...other
   },
   ref,
@@ -216,20 +188,13 @@ export const Button = forwardRef<
     ...other,
   }
 
-  // We need everything in elements for proper flexing 💪
-  const updatedChildren = ReactChildren.map(children, (child) =>
-    typeof child !== 'object' ? (
-      <ButtonCenterContent>{child}</ButtonCenterContent>
-    ) : (
-      child
-    ),
-  )
-
   return (
     <ButtonBase {...buttonProps}>
-      <ButtonInner>{updatedChildren}</ButtonInner>
+      {fullWidth ? (
+        <ButtonFullWidth>{children}</ButtonFullWidth>
+      ) : (
+        <ButtonInner>{children}</ButtonInner>
+      )}
     </ButtonBase>
   )
 })
-
-// Button.displayName = 'Button'
